+++ src/audio_metadata/formats/vorbis.py
@@ -1,4 +1,5 @@
 __all__ = [
+	'VorbisComment',
 	'VorbisComments',
 	'VorbisPicture',
 ]
@@ -6,12 +7,36 @@ __all__ = [
 import struct
 from collections import defaultdict
 
+from attr import attrib, attrs
+from tbm_utils import AttrMapping
+
 from .models import (
 	Picture,
 	Tags,
 )
 from .tables import ID3PictureType
+from ..exceptions import InvalidComment
 from ..utils import datareader
+
+
+@attrs(repr=False)
+class VorbisComment(AttrMapping):
+	name = attrib(converter=lambda n: n.lower())
+	value = attrib()
+
+	@datareader
+	@classmethod
+	def load(cls, data):
+		length = struct.unpack('I', data.read(4))[0]
+		comment = data.read(length).decode('utf-8', 'replace')
+
+		# TODO: Exception.
+		if '=' not in comment:  # pragma: nobranch
+			raise InvalidComment("Vorbis comment must contain an '='.")
+
+		name, value = comment.split('=', 1)
+
+		return cls(name, value)
 
 
 # TODO: Number frames.
@@ -26,13 +51,8 @@ class VorbisComments(Tags):
 		fields = defaultdict(list)
 
 		for _ in range(num_comments):
-			length = struct.unpack('I', data.read(4))[0]
-			comment = data.read(length).decode('utf-8', 'replace')
-
-			if '=' in comment:  # pragma: nobranch
-				field, value = comment.split('=', 1)
-
-				fields[field.lower()].append(value)
+			comment = VorbisComment.load(data)
+			fields[comment.name].append(comment.value)
 
 		return cls(**fields, _vendor=vendor)