+++ CHANGELOG.md
@@ -11,6 +11,8 @@ This project adheres to [Semantic Versioning](https://semver.org).
 ### Added
 
 * ID3 version as attribute on ``ID3v2Frames``.
+* ``VorbisComment``.
+* ``InvalidComment`` exception.
 
 ### Changed